decouple lighter with core (1): move files to core
