+++ app/packages/core/src/components/Modal/Lighter/SharedCanvas.ts
