+++ app/packages/core/src/components/Modal/Lighter/useLighterTooltipEventHandler.ts
@@ -5,8 +5,8 @@
 import * as fos from "@fiftyone/state";
 import { useEffect } from "react";
 import { useRecoilCallback } from "recoil";
-import { LIGHTER_EVENTS, Scene2D } from "../index";
-import type { Hoverable } from "../types";
+import { LIGHTER_EVENTS, Scene2D } from "@fiftyone/lighter";
+import type { Hoverable } from "@fiftyone/lighter/src/types";
 
 /**
  * Hook that handles tooltip events for lighter overlays.
