+++ app/packages/core/src/components/Modal/ModalLooker.tsx
@@ -4,8 +4,8 @@ import { isNativeMediaType } from "@fiftyone/looker/src/util";
 import * as fos from "@fiftyone/state";
 import React, { useMemo } from "react";
 import { useRecoilCallback, useRecoilValue } from "recoil";
-import { LighterSampleRenderer } from "../../../../lighter/src";
 import { ImaVidLookerReact } from "./ImaVidLooker";
+import { LighterSampleRenderer } from "./Lighter/LighterSampleRenderer";
 import { MetadataLooker } from "./MetadataLooker";
 import { VideoLookerReact } from "./VideoLooker";
 import useLooker from "./use-looker";
