+++ app/packages/lighter/src/react/index.ts
@@ -2,14 +2,7 @@
  * Copyright 2017-2025, Voxel51, Inc.
  */
 
-// React components
-export { LighterSampleRenderer } from "./LighterSampleRenderer";
-export type { LighterSampleRendererProps } from "./LighterSampleRenderer";
-
 // React hooks
 export { useLighter } from "./useLighter";
 export { useLighterSetupWithPixi } from "./useLighterSetup";
 export type { LighterOptions } from "./useLighterSetup";
-export { useLighterTooltipEventHandler } from "./useLighterTooltipEventHandler";
-export { usePixiResourceLoader } from "./usePixiResourceLoader";
-export { useSceneSelectionState } from "./useSceneSelectionState";
